Dev: move ControlsOverlay to new file
