+++ charts/mapCharts/MapTab.tsx
@@ -21,7 +21,7 @@ import { easeCubic } from "d3-ease"
 import { ChartLayout, ChartLayoutView } from "charts/core/ChartLayout"
 import { ChartView } from "charts/core/ChartView"
 import { LoadingOverlay } from "charts/core/LoadingOverlay"
-import { ControlsOverlay } from "charts/controls/Controls"
+import { ControlsOverlay } from "charts/controls/ControlsOverlay"
 import { MapTooltip } from "./MapTooltip"
 import { ProjectionChooser } from "./ProjectionChooser"
 import { ColorScale } from "charts/color/ColorScale"
